Added the datasets
Each dataset is described in the descr folder and is taken from the NASA PCoE website. There are also two scripts, base.py and data_tools.py, where base.py cna load the datasets and data_tools can do small modifications to it.

The data tools are currently not in the __init__.py. Will do so when more functionality added to data_tools.py and possibly more generalized.