RUL Estimation v0.01
This repository contains the planned and previously used methods for
Remaining Useful Life (RUL) estimations for the PHM and C-MAPSS datasets
available on the NASA PCoE Repository page.

The hope is to update this methods with newer methodologies and convert it to a fully Python project (as Torch & Lua was used previously). This is also due to the rushed and hacked files used to pipeline the preprocessing and datasets from Python to Lua in a short amount of time.

The committed files are planned projects and existing code that are to be updated for later research.